week-5/task_3/requirements.txt
#accelerate
#bitsandbytes
google-generativeai
langsmith
matplotlib
numpy
openai
pillow
python-dotenv
requests
streamlit
torch
tqdm
transformers
yfinance